+++ BackendBench/backends/llm.py
@@ -16,24 +16,24 @@ from BackendBench.eval import _allclose
 import torch
 
 from .base import Backend
-from BackendBench.llm_client import LLMClient
+from BackendBench.llm_client import LLMKernelGenerator
 
 logger = logging.getLogger(__name__)
 
 
-class LLMRelayBackend(Backend):
+class LLMBackend(Backend):
     """
     Backend that uses LLMKernelGenerator to communicate with local plugboard server or direct Anthropic API.
     """
 
-    def __init__(self, model: str, llm_client: LLMClient) -> None:
-        super().__init__("llm-relay")
+    def __init__(self, model: str, llm_client: LLMKernelGenerator) -> None:
+        super().__init__("llm")
         self.compiled_kernels: Dict[str, Callable] = {}
         self.model = model
         self.llm_client = llm_client
         # Create generated_kernels directory
         timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-        self.kernels_dir = f"generated_kernels/llm_relay_run_{timestamp}"
+        self.kernels_dir = f"generated_kernels/llm_run_{timestamp}"
         os.makedirs(self.kernels_dir, exist_ok=True)
         server_description = llm_client.readme_server_description
         setup_section = llm_client.readme_setup_section
@@ -42,13 +42,13 @@ class LLMRelayBackend(Backend):
         readme_path = os.path.join(self.kernels_dir, "README.md")
         with open(readme_path, "w") as f:
             f.write(
-                f"""# Generated Kernels - LLM Relay - {timestamp}
+                f"""# Generated Kernels - LLM - {timestamp}
 
-This directory contains PyTorch/Triton kernels generated by the LLM Relay Backend.
+This directory contains PyTorch/Triton kernels generated by the LLM Backend.
 
 ## Run Info
 - Timestamp: {timestamp}
-- Backend: LLM Relay
+- Backend: LLM
 - Model: {model}
 - Server: {server_description}
 
@@ -65,7 +65,7 @@ You can inspect these files to debug kernel generation, manually test implementa
 """
             )
 
-        logger.info(f"Saving LLM Relay generated kernels to: {self.kernels_dir}")
+        logger.info(f"Saving LLM generated kernels to: {self.kernels_dir}")
 
     def compile_kernel_from_string(
         self, kernel_code: str, op_name: str, attempt: int = 1
